Trying to train a linear classifier

+++ moco/.ipynb_checkpoints/loader-checkpoint.py
@@ -64,4 +64,4 @@ class RandomDropBands(object):
         chosen_bands = np.random.choice([*range(x.shape[0])], num_bands, replace=False)
         band_mask = torch.stack([torch.ones(x.shape[1:]) if i in chosen_bands else torch.zeros(x.shape[1:]) for i in range(x.shape[0])])
         masked_image = x * band_mask
-        return masked_image.half()
+        return masked_image.float()
